delete

+++ app/Http/Controllers/Api/TodoApiController.php
@@ -64,10 +64,11 @@ class TodoApiController extends BaseController
 
     public function delete(Request $request)
     {
-        $item = Todo::find($request->input('todo_id'));
-//        if (!$item) {
-//            return $this->sendError("Todo not found");
-//        }
+        $id = $request->input('id');
+        $item = Todo::find($id);
+        if (!$item) {
+            return $this->sendError("Todo not found");
+        }
         $item->delete();
         return $this->sendResponse($item->toArray(), "Todo deleted", "todo");
     }
